pyproject.toml
[project]
name = "a-bose"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "pillow>=12.1.0",
    "pytesseract>=0.3.13",
]
